Beheer & Setting and Login screen Functionality Users now only see their registered devices.
Users can login using their username and password.

In the menu screen you can now select a device of witch you want to view the values or change the min-max values.
If no devices are registered to the user it will show no devices in the device selector and this will be disabled.
The sensor overview button disableProperty is bound to the combobox disableProperty so that button will also be disabled.
The settings button is still available but will open an empty page. this will later allow to add a new device.

In both the Beheer and overview you can see the current selected device in the top left.